Fix some of the warnings in #4013

+++ libr/debug/p/debug_native.c
@@ -207,7 +207,7 @@ static int r_debug_native_continue (RDebug *dbg, int pid, int tid, int sig) {
 #elif __BSD__
 	void *data = (void*)(size_t)((sig != -1) ? sig : dbg->reason.signum);
 	ut64 pc = r_debug_reg_get (dbg, "pc");
-	return ptrace (PTRACE_CONT, pid, (void*)(size_t)pc, (int)data) == 0;
+	return ptrace (PTRACE_CONT, pid, (void*)(size_t)pc, (int)(size_t)data) == 0;
 #elif __CYGWIN__
 	#warning "r_debug_native_continue not supported on this platform"
 	return -1;
